requirements.txt
spotipy
instagrapi
python-dotenv